Add a product tax (Asked from cj.com)
If the value is empty, there is no tax on the products.
If the value is a number, it is a percentage (between 0% and 100%). The calculation will be: price * (1/(1+value)).
If the value is not empty and is not a number, then it retrieves the value of the specified attribute. The calculation will be: price * (1 - itemValueTaxField / price)